undo move of reader_manager
* easier to review
